move provider registry and resolution into providers/config.ts
